+++ src/models/hog1p_3d_model.h
@@ -21,19 +21,19 @@ namespace hog3d_cme {
 
     // Function to constraint the shape of the FSP
     void  lhs_constr(PetscInt num_species, PetscInt num_constrs, PetscInt num_states, PetscInt *states,
-                       double *vals){
+                       int *vals){
 
         for (int i{0}; i < num_states; ++i){
-            vals[i*num_constrs] = double(states[num_species*i]);
-            vals[i*num_constrs + 1] = double(states[num_species*i+1]);
-            vals[i*num_constrs + 2] = double(states[num_species*i+2]);
-            vals[i*num_constrs + 3] = double(states[num_species*i]==0)*(double(states[num_species*i+1]) + double(states[num_species*i+2]));
-            vals[i*num_constrs + 4] = double(states[num_species*i]==1)*(double(states[num_species*i+1]) + double(states[num_species*i+2]));
-            vals[i*num_constrs + 5] = double(states[num_species*i]==2)*(double(states[num_species*i+1]) + double(states[num_species*i+2]));
-            vals[i*num_constrs + 6] = double(states[num_species*i]==3)*(double(states[num_species*i+1]) + double(states[num_species*i+2]));;
+            vals[i*num_constrs] = states[num_species*i];
+            vals[i*num_constrs + 1] = states[num_species*i+1];
+            vals[i*num_constrs + 2] = states[num_species*i+2];
+            vals[i*num_constrs + 3] = (states[num_species*i]==0)*((states[num_species*i+1]) + (states[num_species*i+2]));
+            vals[i*num_constrs + 4] = (states[num_species*i]==1)*((states[num_species*i+1]) + (states[num_species*i+2]));
+            vals[i*num_constrs + 5] = (states[num_species*i]==2)*((states[num_species*i+1]) + (states[num_species*i+2]));
+            vals[i*num_constrs + 6] = (states[num_species*i]==3)*((states[num_species*i+1]) + (states[num_species*i+2]));;
         }
     }
-    arma::Row<double> rhs_constr{3.0, 4.0, 4.0, 1.0, 10.0, 10.0, 10.0};
+    arma::Row<int> rhs_constr{3, 4, 4, 1, 10, 10, 10};
     arma::Row<double> expansion_factors{0.0, 0.5, 0.5, 0.5,0.5, 0.5, 0.5, 0.5,0.5, 0.5, 0.5};
 
 // propensity function
